breaking apart composables
