+++ resources/js/pages/reports/SchedulingReport/InstructorTable.vue
@@ -19,10 +19,6 @@
           }"
         >
           {{ term.name }}
-          <Spinner
-            v-if="isTermLoaded(term.id)"
-            class="tw-text-neutral-300 tw-h-4 tw-w-4"
-          />
         </Th>
       </tr>
     </template>
@@ -41,25 +37,16 @@
 <script setup lang="ts">
 import { Table, Th } from "@/components/Table";
 import InstructorTableRow from "./InstructorTableRow.vue";
-import { Term, Instructor, Leave, Course, LoadState } from "@/types";
+import { Term, Instructor, Leave, Course } from "@/types";
 import Spinner from "@/components/Spinner.vue";
 
-const props = defineProps<{
+defineProps<{
   terms: Term[];
   instructors: Instructor[];
   currentTerm: Term | null;
-  termLoadState: Map<Term["id"], LoadState>;
   getLeavesForInstructorPerTerm: (instructorId: number) => Leave[][];
   getCoursesForInstructorPerTerm: (instructorId: number) => Course[][];
   search: string;
 }>();
-
-function isTermLoaded(termId: number) {
-  const termLoadState = props.termLoadState.get(termId);
-  if (!termLoadState) {
-    throw new Error(`Term load state not found for term ${termId}`);
-  }
-  return ["idle", "loading"].includes(termLoadState);
-}
 </script>
 <style scoped></style>
